refactor to inside model
